create AlloyFeaturesTest
